+++ android/app/src/main/java/com/daedan/festabook/data/datasource/remote/faq/FAQDataSource.kt
@@ -1,7 +1,7 @@
 package com.daedan.festabook.data.datasource.remote.faq
 
 import com.daedan.festabook.data.datasource.remote.ApiResult
-import com.daedan.festabook.data.model.response.FAQResponse
+import com.daedan.festabook.data.model.response.faq.FAQResponse
 
 interface FAQDataSource {
     suspend fun fetchAllFAQs(): ApiResult<List<FAQResponse>>
