+++ android/app/src/main/java/com/daedan/festabook/data/datasource/remote/faq/FAQDataSourceImpl.kt
@@ -1,7 +1,7 @@
 package com.daedan.festabook.data.datasource.remote.faq
 
 import com.daedan.festabook.data.datasource.remote.ApiResult
-import com.daedan.festabook.data.model.response.FAQResponse
+import com.daedan.festabook.data.model.response.faq.FAQResponse
 import com.daedan.festabook.data.service.FAQService
 
 class FAQDataSourceImpl(
