+++ android/app/src/main/java/com/daedan/festabook/data/repository/FAQRepositoryImpl.kt
@@ -1,7 +1,7 @@
 package com.daedan.festabook.data.repository
 
 import com.daedan.festabook.data.datasource.remote.faq.FAQDataSource
-import com.daedan.festabook.data.model.response.toDomain
+import com.daedan.festabook.data.model.response.faq.toDomain
 import com.daedan.festabook.data.util.toResult
 import com.daedan.festabook.domain.model.FAQItem
 import com.daedan.festabook.domain.repository.FAQRepository
