+++ android/app/src/main/java/com/daedan/festabook/data/service/FAQService.kt
@@ -1,6 +1,6 @@
 package com.daedan.festabook.data.service
 
-import com.daedan.festabook.data.model.response.FAQResponse
+import com.daedan.festabook.data.model.response.faq.FAQResponse
 import retrofit2.Response
 import retrofit2.http.GET
